# perf: Remove custom paging option from main page controller