## Motivation

Slow query execution at https://eth-sepolia.k8s-dev.blockscout.com/api/v2/main-page/transactions/watchlist

## Changelog

```
SELECT t0.*, (SELECT transaction_hash FROM token_transfers WHERE transaction_hash = t0.hash LIMIT 1) IS NOT NULL, b1.* 
FROM transactions AS t0 
INNER JOIN blocks AS b1 ON b1.hash = t0.block_hash 
WHERE (NOT (t0.block_number IS NULL)) 
    AND (t0.to_address_hash = ANY (ARRAY['\x91be515287a961a8d09fd4ca3dfb5e76ed01575e'::BYTEA, '\x993a0f3653887078215914badcf039263293add9'::BYTEA, '\x13cb6ae34a13a0977f4d7101ebc24b87bb23f0d5'::BYTEA])) 
ORDER BY t0.block_number DESC, t0.index DESC 
LIMIT 6;
```

never finishes.

At the same time,

```
EXPLAIN ANALYZE SELECT t0.*, (SELECT transaction_hash FROM token_transfers WHERE transaction_hash = t0.hash LIMIT 1) IS NOT NULL
FROM transactions AS t0 
WHERE (NOT (t0.block_number IS NULL)) 
    AND (t0.to_address_hash = ANY (ARRAY['\x91be515287a961a8d09fd4ca3dfb5e76ed01575e'::BYTEA, '\x993a0f3653887078215914badcf039263293add9'::BYTEA, '\x13cb6ae34a13a0977f4
d7101ebc24b87bb23f0d5'::BYTEA])) 
ORDER BY t0.block_number DESC, t0.index DESC 
LIMIT 51;
                                                                                                                       QUERY PLAN                                                                                                                        
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=287164.10..287213.96 rows=51 width=606) (actual time=139.101..160.310 rows=51 loops=1)
   ->  Result  (cost=287164.10..356475.21 rows=70907 width=606) (actual time=128.803..150.000 rows=51 loops=1)
         ->  Sort  (cost=287164.10..287341.37 rows=70907 width=605) (actual time=128.195..128.230 rows=51 loops=1)
               Sort Key: t0.block_number DESC, t0.index DESC
               Sort Method: top-N heapsort  Memory: 70kB
               ->  Bitmap Heap Scan on transactions t0  (cost=8060.90..284798.50 rows=70907 width=605) (actual time=3.800..127.272 rows=659 loops=1)
                     Recheck Cond: ((to_address_hash = ANY ('{"\\x91be515287a961a8d09fd4ca3dfb5e76ed01575e","\\x993a0f3653887078215914badcf039263293add9","\\x13cb6ae34a13a0977f4d7101ebc24b87bb23f0d5"}'::bytea[])) AND (block_number IS NOT NULL))
                     Heap Blocks: exact=658
                     ->  Bitmap Index Scan on transactions_to_address_hash_with_pending_index_asc  (cost=0.00..8043.17 rows=70907 width=0) (actual time=3.507..3.507 rows=660 loops=1)
                           Index Cond: ((to_address_hash = ANY ('{"\\x91be515287a961a8d09fd4ca3dfb5e76ed01575e","\\x993a0f3653887078215914badcf039263293add9","\\x13cb6ae34a13a0977f4d7101ebc24b87bb23f0d5"}'::bytea[])) AND (block_number IS NOT NULL))
         SubPlan 1
           ->  Limit  (cost=0.70..0.96 rows=1 width=33) (actual time=0.424..0.424 rows=0 loops=51)
                 ->  Index Only Scan using token_transfers_transaction_hash_log_index_index on token_transfers  (cost=0.70..3889.47 rows=14675 width=33) (actual time=0.423..0.423 rows=0 loops=51)
                       Index Cond: (transaction_hash = t0.hash)
                       Heap Fetches: 0
 Planning Time: 0.229 ms
 JIT:
   Functions: 13
   Options: Inlining false, Optimization false, Expressions true, Deforming true
   Timing: Generation 1.013 ms, Inlining 0.000 ms, Optimization 0.831 ms, Emission 9.507 ms, Total 11.351 ms
 Execution Time: 161.547 ms
(21 rows)
```

executes almost immediately.

## Checklist for your Pull Request (PR)

- [ ] If I added new functionality, I added tests covering it.
- [ ] If I fixed a bug, I added a regression test to prevent the bug from silently reappearing again.
- [ ] I checked whether I should update the docs and did so by submitting a PR to [docs repository](https://github.com/blockscout/docs).
- [ ] If I added/changed/removed ENV var, I submitted a PR to [docs repository](https://github.com/blockscout/docs) to update the list of [env vars](https://github.com/blockscout/docs/blob/master/setup/env-variables/README.md) and I updated the version to `master` in the Version column. If I removed variable, I added it to [Deprecated ENV Variables](https://github.com/blockscout/docs/blob/master/setup/env-variables/deprecated-env-variables/README.md) page. After merging docs PR, changes will be reflected in these [pages](https://docs.blockscout.com/setup/env-variables).
- [ ] If I added new DB indices, I checked, that they are not redundant, with PGHero or other tools.
- [ ] If I added/removed chain type, I modified the Github CI matrix and PR labels accordingly.


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **Refactor**
  - Removed fixed page size limit for transactions on the main page API.
- **Tests**
  - Expanded recent transactions tested from 6 to 50+ to better match real data volumes.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->